requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
blis==0.7.6
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cymem==2.0.6
Cython==0.29.23
debugpy==1.5.1
decorator==5.1.1
dill==0.3.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.2
filelock==3.6.0
gensim==4.1.2
huggingface-hub==0.4.0
idna==3.3
ipykernel==6.9.1
ipython==8.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pickleshare==0.7.5
platformdirs==2.5.0
preshed==3.0.6
prompt-toolkit==3.0.28
pure-eval==0.2.2
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
thinc==8.0.13
tokenizers==0.11.5
tomli==2.0.1
torch==1.10.2
torchsummary==1.5.1
torchtext==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
